# Conflicts:
#	GUI-dev/package.json
#	GUI-dev/src/app/@theme/components/footer/footer.component.ts
#	GUI-dev/src/app/@theme/providers/backend-server/gateway.ts
#	GUI-dev/src/app/@theme/providers/backend-server/serverconfig.ts
#	GUI-dev/src/app/allProcesses.service.ts
#	GUI-dev/src/app/app.module.ts
#	GUI-dev/src/app/eventLogger.service.ts
#	GUI-dev/src/app/pages/admin/admin.module.ts
#	GUI-dev/src/app/pages/admin/admin.routing.ts
#	GUI-dev/src/app/pages/admin/components/pay-as-you-go/pay-as-you-go.component.html
#	GUI-dev/src/app/pages/admin/components/pay-as-you-go/pay-as-you-go.component.scss
#	GUI-dev/src/app/pages/admin/components/pay-as-you-go/pay-as-you-go.component.ts
#	GUI-dev/src/app/pages/pages-menu.ts
#	GUI-dev/src/app/pages/pages-routing.module.ts
#	GUI-dev/src/app/pages/pages.module.ts
#	GUI-dev/src/app/pipes/commaDate.pipe.ts
#	GUI-dev/src/app/pipes/engineDate.pipe.ts
#	GUI-dev/src/app/pipes/pipes.module.ts
#	GUI-dev/src/models/models.ts
#	Gateway/src/main/java/at/fhjoanneum/ippr/gateway/api/controller/ProcessModelStorageGatewayController.java
#	Gateway/src/main/java/at/fhjoanneum/ippr/gateway/api/services/impl/ProcessModelStorageCallerImpl.java
#	Persistence/src/main/java/at/fhjoanneum/ippr/persistence/entities/model/payasyougo/PayAsYouGoImpl.java
#	ProcessModelStorage/src/main/java/at/fhjoanneum/ippr/pmstorage/controller/ProcessModelStorageController.java
#	ProcessModelStorage/src/main/java/at/fhjoanneum/ippr/pmstorage/repositories/PayAsYouGoRepository.java
#	ProcessModelStorage/src/main/java/at/fhjoanneum/ippr/pmstorage/services/ProcessModelService.java
#	ProcessModelStorage/src/main/java/at/fhjoanneum/ippr/pmstorage/services/impl/ProcessModelServiceImpl.java
#	SpringCommons/src/main/java/at/fhjoanneum/ippr/commons/dto/payasyougo/PayAsYouGoDTO.java

+++ ProcessModelStorage/src/main/java/at/fhjoanneum/ippr/pmstorage/repositories/PayAsYouGoRepository.java
@@ -6,7 +6,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.Modifying;
-
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -28,7 +27,8 @@ public interface PayAsYouGoRepository extends PagingAndSortingRepository<PayAsYo
   public int updatePayAsYouGoByPiId(@Param("datetime_end") LocalDateTime datetime_end,
                                                @Param("pi_id") Long pi_id);
 
+
   @Query(value = "SELECT * FROM pay_as_you_go p WHERE p.org_id = :org_id", nativeQuery = true)
-  public List<PayAsYouGoImpl> findPayAsYouGoByOrgId(@Param("org_id") Long org_id);
+  public List<PayAsYouGoImpl> findPayAsYouGoByOrgId(@Param("org_id") int org_id);
 
 }
